Further villager changes
Moved scroller and buttons a bit to make gui look better, before buttons were cut off by one pixel at the top and the scroller was too close to the right.
The only problem is that scroller extends too much (by two screen pixels) at the bottom and I didn't figure out how to change it. I'll add the visualization to the "Suggestions".