Dispenser/Dropper animation, exp bar fixes
If player is on mobile, HUD gets wider because of "inventory_touch_button" (button that opens inventory, with three elipses), so the exp bar was too little.
The similar thing happens to horse jump bar, but I didn't figure out how to fix it, and it don't look too bad.