enchant toggles
I changed it so the enchanting table toggles require a double-press to select on mobile. I did this because when a mobile player tries to hover to view the enchant info, it automatically selects the button thus this should fix that.

will be doing some more things with the code soon to hopefully improve performance